pyproject.toml
# TensorFlow VS Pytorch

[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# Project settings -----------------------------------------------------------------------------------------------------
[project]
name = "tfvspt"
dynamic = ["version"]
description = "Comparing TensorFlow and Pytorch"
readme = "README.md"
requires-python = ">=3.10"

# Required dependencies ------------------------------------------------------------------------------------------------
dependencies = [
    "tqdm >= 4.67.1",
    "torch >= 2.6.0",
    "PyYAML >= 6.0.2",
    "notebook >= 7.3.3",
    "pydantic >= 2.10.6",
    "torchinfo >= 1.8.0",
    "tensorflow >= 2.19.0",
    "matplotlib >= 3.10.1",
    "torchvision >= 0.21.0",
]

# Optional dependencies ------------------------------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "yapf",
    "isort",
    "pre-commit",
]

[project.urls]
"Source" = "https://github.com/infocusp/tf_pt_benchmarking"

[project.scripts]
tfvspt = "tfvspt.main:entrypoint"
downloadcifar100 = "tfvspt.download_data:entrypoint"

# Tools settings -------------------------------------------------------------------------------------------------------
[tool.setuptools]  # configuration specific to the `setuptools` build backend.
packages = { find = { where = ["."], include = ["tfvspt", "tfvspt.*"] } }

[tool.setuptools.dynamic]
version = { attr = "tfvspt.__version__" }
